+++ src/main/java/com/exaroton/api/server/config/ConfigOptionTypeAdapterFactory.java
@@ -1,0 +1,17 @@
+package com.exaroton.api.server.config;
+
+import com.google.gson.Gson;
+import com.google.gson.TypeAdapter;
+import com.google.gson.TypeAdapterFactory;
+import com.google.gson.reflect.TypeToken;
+
+public class ConfigOptionTypeAdapterFactory implements TypeAdapterFactory {
+    @Override
+    public <T> TypeAdapter<T> create(Gson gson, TypeToken<T> type) {
+        if (ConfigOption.class.isAssignableFrom(type.getRawType())) {
+            //noinspection unchecked
+            return (TypeAdapter<T>) new ConfigOptionTypeAdapter();
+        }
+        return null;
+    }
+}