+++ src/main/java/com/exaroton/api/server/config/options/BooleanConfigOption.java
@@ -2,22 +2,12 @@ package com.exaroton.api.server.config.options;
 
 import com.exaroton.api.server.config.ConfigOption;
 import com.exaroton.api.server.config.OptionType;
-import org.jetbrains.annotations.Nullable;
 
-public class BooleanConfigOption extends ConfigOption {
-    private Boolean value;
-    public BooleanConfigOption(String key, Boolean value, String label, String[] options) {
-        super(key, label, OptionType.BOOLEAN, options);
-        this.value = value;
-    }
-
-    public BooleanConfigOption setValue(Boolean value) {
-        this.value = value;
-        return this;
-    }
-
-    @Override
-    public @Nullable Boolean getValue() {
-        return value;
+/**
+ * A boolean option
+ */
+public class BooleanConfigOption extends ConfigOption<Boolean> {
+    public BooleanConfigOption(String key, Boolean value, String label) {
+        super(key, value, label, OptionType.BOOLEAN);
     }
 }