+++ src/main/java/com/exaroton/api/server/config/options/SelectConfigOption.java
@@ -1,24 +1,15 @@
 package com.exaroton.api.server.config.options;
 
-import com.exaroton.api.server.config.ConfigOption;
 import com.exaroton.api.server.config.OptionType;
-import org.jetbrains.annotations.Nullable;
 
-public class SelectConfigOption extends ConfigOption {
-    private String value;
+import java.util.Set;
 
-    public SelectConfigOption(String key, String value, String label, String[] options) {
-        super(key, label, OptionType.SELECT, options);
-        this.value = value;
-    }
+/**
+ * A select option
+ */
+public class SelectConfigOption extends BaseSelectOption<String> {
 
-    public SelectConfigOption setValue(String value) {
-        this.value = value;
-        return this;
-    }
-
-    @Override
-    public @Nullable String getValue() {
-        return value;
+    public SelectConfigOption(String key, String value, String label, Set<String> options) {
+        super(key, value, label, OptionType.SELECT, options);
     }
 }