+++ src/test/java/ConfigTest.java
@@ -1,8 +1,8 @@
 import com.exaroton.api.APIException;
-import com.exaroton.api.server.Server;
 import com.exaroton.api.server.ServerFile;
 import com.exaroton.api.server.config.ConfigOption;
 import com.exaroton.api.server.config.ServerConfig;
+import com.exaroton.api.server.config.options.SelectConfigOption;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -14,11 +14,15 @@ public class ConfigTest extends APIClientTest {
         ServerConfig config = serverProperties.getConfig();
         Assertions.assertNotNull(config);
 
-        ConfigOption gamemodeOption = config.getOption("gamemode");
-        Assertions.assertNotNull(gamemodeOption);
-        Assertions.assertEquals("Gamemode", gamemodeOption.getLabel());
-        Assertions.assertNotNull(gamemodeOption.getValue());
-        Assertions.assertNotNull(gamemodeOption.getOptions());
-        Assertions.assertTrue(gamemodeOption.getOptions().length > 3);
+        ConfigOption<?> gamemodeOption = config.getOption("gamemode");
+        Assertions.assertInstanceOf(SelectConfigOption.class, gamemodeOption);
+        SelectConfigOption select = (SelectConfigOption) gamemodeOption;
+
+        Assertions.assertNotNull(select);
+        Assertions.assertEquals("Gamemode", select.getLabel());
+        Assertions.assertNotNull(select.getValue());
+
+        Assertions.assertNotNull(select.getOptions());
+        Assertions.assertTrue(select.getOptions().size() > 3);
     }
 }