+++ CHANGELOG.md
@@ -8,6 +8,7 @@ from this library, you must include an SLF4J implementation in your project.
 
 ### Other
 - Arrays have been Replaced by Collection's in almost all places
+- Config options now return a generic type instead of `Object`
 - Removed `ExarotonClient#getGson()` and `WebsocketClient#getGson()`
 - Renamed `Server#setClient` to `Server#init` and added parameter `gson`